core/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
celery
redis
sqlalchemy
asyncpg
alembic
httpx
openai
gspread
google-auth